+++ bin/run.dart
@@ -2,7 +2,6 @@
 
 import 'package:h4/create.dart';
 import 'package:h4/src/create_error.dart';
-import 'package:h4/src/logger.dart';
 
 void main(List<String> arguments) async {
   var app = createApp();
@@ -19,7 +18,10 @@ void main(List<String> arguments) async {
       (error, stack, event) => print('Error occured at ${event?.path}'));
 
   router.get<bool>("/25/**", (event) => true);
-  router.get("/", (event) => throw ("HSHAHAHAH"));
+  router.get("/:id", (event) {
+    print('param: ${event.params["id"]}');
+    return 'Hey';
+  });
 
   router.post("/vamos", (event) async {
     await Future.delayed(Duration(milliseconds: 100));