+++ lib/src/event.dart
@@ -6,7 +6,7 @@ import 'dart:io';
 /// The `H4Event` class encapsulates an HTTP request and provides methods and properties
 /// to interact with the request and generate the appropriate response.
 class H4Event {
-  Map<String, String> _params;
+  Map<String, String> params;
 
   /// The HTTP request that triggered the event.
   ///
@@ -16,7 +16,7 @@ class H4Event {
   /// Tells us whether the event has been handled and a response has been generated.
   bool handled = false;
 
-  H4Event(this._request) : _params = {};
+  H4Event(this._request) : params = {};
 
   String get path => _request.uri.path;
 
@@ -29,7 +29,7 @@ class H4Event {
 
   /// Sets the event parameters to the provided [params] map.
   set eventParams(Map<String, String> params) {
-    _params = params;
+    this.params = params;
   }
 
   /// The status message associated with the HTTP response status code.